+++ com.unity.netcode.gameobjects/Runtime/Transports/SinglePlayer.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f4bf40cad73d0464e8656d8c7c2062db
+guid: 2673df68d03763c43bcce03e872b91e7
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
