+++ com.unity.netcode.gameobjects/Tests/Runtime/Transports/SinglePlayerTransportTests.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using NUnit.Framework;
+using Unity.Netcode.Transports.SinglePlayer;
 using Unity.Netcode.TestHelpers.Runtime;
 using UnityEngine;
 using UnityEngine.TestTools;
